Code Refactoring Fixed broken tests
