+++ opensrp-chw/src/main/res/layout/activity_guide_books_counseling.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context=".fragment.JobAidsGuideBooksTutorialsActivity">
+    tools:context=".activity.GuideBooksTutorialsActivity">
 
     <com.github.barteksc.pdfviewer.PDFView
         android:id="@+id/pdfView"
